requirements.txt
click==6.2
coverage==4.0.3
django==1.9.2
ipdb==0.8.1
ipython==4.1.1
pytest-cov==2.2.1
python-slugify==1.2.0
PyYAML==3.11
six==1.10.0
twine==1.6.5
